Finished splitting up core
