+++ frontpy/core/back_end/__init__.py
@@ -18,9 +18,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-import sys
 from __custom_import__ import prepare_import
-prepare_import("front_end","back_end")
-from back_end import *
-from front_end import *
-from exceptions import *
+prepare_import()
+from Core_http_process import Core_http_process
+from functions import serve,run_app
