+++ frontpy/core/constants.py
@@ -18,9 +18,11 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-import sys
-from __custom_import__ import prepare_import
-prepare_import("front_end","back_end")
-from back_end import *
-from front_end import *
-from exceptions import *
+
+JQUERY_3_2_1_MIN_JS = "234c8514654bb7ed8a60ea905b6f98f0"
+BOOTSTRAP_MIN_JS = "13b2a30e265e18a6fd0792cc3fd7a09c"
+SCRIPT_JS = "9a9569e9d73f33740eada95275da7f30"
+BOOTSTRAP_CSS = "e3202aea761d3d587dfcfc43c6982565"
+
+INVALID_INSERTION_MESSAGE = "You can't use '%s', please try some object from the framework library"
+AVALIABLE_FORMATS = ("jpg","png","svg","map","ico","js","css") #Formats of files that can be loaded automatically from pages/media folder
